refactor: centralize the organizations dropdown logic
